+++ src/main/java/com/spotify/docker/client/DockerConfigReader.java
@@ -21,33 +21,21 @@
 package com.spotify.docker.client;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.common.base.Strings.isNullOrEmpty;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.google.common.base.Preconditions;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableCollection;
+import com.spotify.docker.client.messages.DockerCredentialHelperAuth;
 import com.spotify.docker.client.messages.RegistryAuth;
-import com.spotify.docker.client.messages.RegistryAuthV2;
 import com.spotify.docker.client.messages.RegistryConfigs;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -56,12 +44,19 @@ public class DockerConfigReader {
   private static final Logger LOG = LoggerFactory.getLogger(DockerConfigReader.class);
 
   private static final ObjectMapper MAPPER = ObjectMapperProvider.objectMapper();
-  private static final String AUTHS_ENTRY = "auths";
-  private static final String CREDS_STORE = "credsStore";
 
-  /** Returns all RegistryConfig instances from the configuration file. */
+
+  /**
+   * Parse the contents of the config file and generate all possible
+   * {@link RegistryAuth}s, which are bundled into a {@link RegistryConfigs} instance.
+   * @param configPath Path to config file.
+   * @return All registry configs that can be generated from the config file
+   * @throws IOException If the file cannot be read, or its JSON cannot be parsed
+   * @deprecated Use {@link #authForAllRegistries(Path)} instead.
+   */
+  @Deprecated
   public RegistryConfigs fromConfig(final Path configPath) throws IOException {
-    return parseDockerConfig(configPath);
+    return authForAllRegistries(configPath);
   }
 
   /**
@@ -69,50 +64,154 @@ public class DockerConfigReader {
    *
    * @throws IllegalArgumentException if the config file does not contain registry auth info for the
    *                                  registry
+   * @deprecated In favor of {@link #authForRegistry(Path, String)}
    */
+  @Deprecated
   public RegistryAuth fromConfig(final Path configPath, final String serverAddress)
       throws IOException {
-    return parseDockerConfig(configPath, serverAddress);
+    return authForRegistry(configPath, serverAddress);
   }
 
   /**
-   * @deprecated do not use - only exists for backwards compatibility. Use {@link #fromConfig(Path)}
-   *     instead.
+   * Return a single RegistryAuth from the default config file.
+   * If there is only one, it'll be that one.
+   *
+   * @return Some registry auth value.
    */
-  @Deprecated
-  public RegistryAuth fromFirstConfig(Path configPath) throws IOException {
-    return parseDockerConfig(configPath, null);
+  public RegistryAuth anyRegistryAuth() throws IOException {
+    return anyRegistryAuth(defaultConfigPath());
   }
 
-  private RegistryAuth parseDockerConfig(final Path configPath, final String serverAddress)
-      throws IOException {
+  /**
+   * Return a single RegistryAuth from the config file.
+   * If there are multiple RegistryAuth entries, which entry is returned from this method
+   * depends on hashing and should not be considered reliable.
+   * If there is only one entry, however, that will be the one returned. This is the
+   * primary use of this method, as a useful way to extract a RegistryAuth during testing.
+   * In that environment, the contents of the config file are known and controlled. In a
+   * production environment, the contents of the config file are less predictable.
+   *
+   * @param configPath Path to the docker config file.
+   * @return Some registry auth value.
+   */
+  @VisibleForTesting
+  RegistryAuth anyRegistryAuth(final Path configPath) throws IOException {
+    final ImmutableCollection<RegistryAuth> registryAuths =
+        authForAllRegistries(configPath).configs().values();
+    return registryAuths.isEmpty()
+        ? RegistryAuth.builder().build()
+        : registryAuths.iterator().next();
+  }
+
+  /**
+   * Parse the contents of the config file and generate all possible
+   * {@link RegistryAuth}s, which are bundled into a {@link RegistryConfigs} instance.
+   * @param configPath Path to config file.
+   * @return All registry auths that can be generated from the config file
+   * @throws IOException If the file cannot be read, or its JSON cannot be parsed
+   */
+  public RegistryConfigs authForAllRegistries(final Path configPath) throws IOException {
     checkNotNull(configPath);
 
-    final Map<String, RegistryAuth> configs = parseDockerConfig(configPath).configs();
+    final DockerConfig config = MAPPER.readValue(configPath.toFile(), DockerConfig.class);
+    if (config == null) {
+      return RegistryConfigs.empty();
+    }
 
-    if (isNullOrEmpty(serverAddress)) {
-      if (configs.isEmpty()) {
-        return RegistryAuth.builder().build();
+    final RegistryConfigs.Builder registryConfigsBuilder = RegistryConfigs.builder();
+
+    final Map<String, String> credsHelpers = config.credsHelpers();
+    final boolean hasCredsHelpers = credsHelpers != null && !credsHelpers.isEmpty();
+    final Map<String, RegistryAuth> auths = config.auths();
+    final boolean hasAuths = auths != null && !auths.isEmpty();
+    final String credsStore = config.credsStore();
+    final boolean hasCredsStore = credsStore != null;
+
+    // First use the credsHelpers, if there are any
+    if (hasCredsHelpers) {
+      for (final Map.Entry<String, String> credsHelpersEntry : credsHelpers.entrySet()) {
+        final String registry = credsHelpersEntry.getKey();
+        final String aCredsStore = credsHelpersEntry.getValue();
+        registryConfigsBuilder.addConfig(registry,
+            authWithCredentialHelper(aCredsStore, registry));
       }
-      LOG.warn("Returning first entry from docker config file - use fromConfig(Path) instead, "
-               + "this behavior is deprecated and will soon be removed");
-      return configs.values().iterator().next();
     }
 
-    if (configs.containsKey(serverAddress)) {
-      return configs.get(serverAddress);
+    // If there are any objects in "auths", they could take two forms.
+    // Older auths will map registry keys to objects with "auth" values, sometimes emails.
+    // Newer auths will map registry keys to empty objects. They expect you
+    // to use the credsStore to authenticate.
+    if (hasAuths) {
+      // We will use this empty RegistryAuth to check for empty auth values
+      final RegistryAuth empty = RegistryAuth.builder().build();
+
+      for (final Map.Entry<String, RegistryAuth> authEntry : auths.entrySet()) {
+        final String registry = authEntry.getKey();
+        final RegistryAuth registryAuth = authEntry.getValue();
+        if (registryAuth == null || registryAuth.equals(empty)) {
+          // We have an empty object. Can we use credsStore?
+          if (hasCredsStore) {
+            registryConfigsBuilder.addConfig(registry,
+                authWithCredentialHelper(credsStore, registry));
+          } // no else clause. If we can't fall back to credsStore, we can't auth.
+        } else {
+          // The auth object isn't empty.
+          // We need to add the registry to its properties, then
+          // add it to the RegistryConfigs
+          registryConfigsBuilder.addConfig(registry,
+              registryAuth.toBuilder().serverAddress(registry).build());
+        }
+      }
     }
 
+    // If there are no credsHelpers or auths or credsStore, then the
+    // config may be in a very old format. There aren't any keys for different
+    // sections. The file is just a map of registries to auths.
+    // In other words, it looks like a RegistryConfigs.
+    // If we can map it to one, we'll return it.
+    if (!(hasAuths || hasCredsHelpers || hasCredsStore)) {
+      try {
+        return MAPPER.readValue(configPath.toFile(), RegistryConfigs.class);
+      } catch (IOException ignored) {
+        // Looks like that failed to parse.
+        // Eat the exception, fall through, and return empty object.
+      }
+    }
+
+    return registryConfigsBuilder.build();
+  }
+
+  /**
+   * Generate {@link RegistryAuth} for the registry.
+   *
+   * @param configPath Path to the docker config file
+   * @param registry Docker registry for which to generate auth
+   * @return The generated authentication object
+   */
+  public RegistryAuth authForRegistry(final Path configPath, final String registry)
+      throws IOException {
+    checkNotNull(configPath);
+    checkNotNull(registry);
+
+    final DockerConfig config = MAPPER.readValue(configPath.toFile(), DockerConfig.class);
+    if (config == null) {
+      return RegistryAuth.builder().build();
+    }
+
+    final RegistryAuth registryAuth = authForRegistry(config, registry);
+    if (registryAuth != null) {
+      return registryAuth;
+    }
     // If the given server address didn't have a protocol try adding a protocol to the address.
     // This handles cases where older versions of Docker included the protocol when writing
     // auth tokens to config.json.
     try {
-      final URI serverAddressUri = new URI(serverAddress);
+      final URI serverAddressUri = new URI(registry);
       if (serverAddressUri.getScheme() == null) {
         for (String proto : Arrays.asList("https://", "http://")) {
-          final String addrWithProto = proto + serverAddress;
-          if (configs.containsKey(addrWithProto)) {
-            return configs.get(addrWithProto);
+          final RegistryAuth protoRegistryAuth = authForRegistry(config, proto + registry);
+          if (protoRegistryAuth != null) {
+            return protoRegistryAuth;
           }
         }
       }
@@ -121,65 +220,25 @@ public class DockerConfigReader {
     }
 
     throw new IllegalArgumentException(
-        "serverAddress=" + serverAddress + " does not appear in config file at " + configPath);
+        "registry \"" + registry + "\" does not appear in config file at " + configPath);
   }
 
-  private RegistryConfigs parseDockerConfig(final Path configPath) throws IOException {
-    checkNotNull(configPath);
+  private RegistryAuth authForRegistry(final DockerConfig config, final String registry)
+      throws IOException {
 
-    ObjectNode authJson = extractAuthJson(configPath);
-
-    if (authJson.has(CREDS_STORE) && authJson.has(AUTHS_ENTRY)) {
-      String credsStore = authJson.get(CREDS_STORE).textValue();
-      Map<String, RegistryAuth> registryAuthMap = new HashMap<>();
-
-      ObjectNode auths = (ObjectNode)authJson.get(AUTHS_ENTRY);
-      Iterator<String> serverIterator = auths.fieldNames();
-
-      while (serverIterator.hasNext()) {
-        String serverAddress = serverIterator.next();
-
-        Process process = Runtime.getRuntime().exec("docker-credential-" + credsStore + " get");
-
-        try (Writer outStreamWriter = new OutputStreamWriter(
-                                        process.getOutputStream(), StandardCharsets.UTF_8)) {
-          try (BufferedWriter writer = new BufferedWriter(outStreamWriter)) {
-
-            writer.write(serverAddress + "\n");
-            writer.flush();
-          }
-        }
-
-        try (InputStreamReader reader = new InputStreamReader(
-                                          process.getInputStream(), StandardCharsets.UTF_8)) {
-          try (BufferedReader input = new BufferedReader(reader)) {
-            String serverAuthDetails = input.readLine();
-            // ErrCredentialsNotFound standardizes the not found error, so every helper returns
-            // the same message and docker can handle it properly.
-            // https://github.com/docker/docker-credential-helpers/blob/19b711cc92fbaa47533646fa8adb457d199c99e1/credentials/error.go#L4-L6
-            if ("credentials not found in native keychain".equals(serverAuthDetails)) {
-              continue;
-            }
-            JsonNode serverAuthNode = MAPPER.readTree(serverAuthDetails);
-            RegistryAuthV2 serverAuth =
-                new RegistryAuthV2(serverAuthNode.get("Username").textValue(),
-                                   serverAuthNode.get("Secret").textValue(),
-                                   serverAuthNode.get("ServerURL").textValue());
-
-            registryAuthMap.put(serverAddress, serverAuth);
-          }
-        }
-      }
-      return RegistryConfigs.create(registryAuthMap);
-    } else if (authJson.has(AUTHS_ENTRY)) {
-      return MAPPER.treeToValue(authJson.get(AUTHS_ENTRY), RegistryConfigs.class);
+    // If the registry shows up in "auths", return it
+    final Map<String, RegistryAuth> auths = config.auths();
+    if (auths != null && auths.get(registry) != null) {
+      return auths.get(registry).toBuilder().serverAddress(registry).build();
     }
 
-    try {
-      return MAPPER.treeToValue(authJson, RegistryConfigs.class);
-    } catch (JsonProcessingException e) {
-      return RegistryConfigs.empty();
+    // Else, we use a credential helper.
+    final String credsStore = getCredentialStore(config, registry);
+    if (credsStore != null) {
+      return authWithCredentialHelper(credsStore, registry);
     }
+
+    return null;
   }
 
   public Path defaultConfigPath() {
@@ -196,20 +255,32 @@ public class DockerConfigReader {
     }
   }
 
-  private ObjectNode extractAuthJson(final Path configPath) throws IOException {
-    final File file = configPath.toFile();
+  /**
+   * Obtain auth using a credential helper.
+   * @param credsStore The name of the credential helper
+   * @param registry The registry for which we need to obtain auth
+   * @return A RegistryAuth object with a username, password, and server.
+   * @throws IOException This method attempts to execute
+   *                     "docker-credential-" + credsStore + " get". If you don't have the
+   *                     proper credential helper installed and on your path, this
+   *                     will fail.
+   */
+  private RegistryAuth authWithCredentialHelper(final String credsStore,
+                                                final String registry) throws IOException {
+    final DockerCredentialHelperAuth dockerCredentialHelperAuth =
+        DockerCredentialHelper.get(credsStore, registry);
+    return dockerCredentialHelperAuth == null ? null : dockerCredentialHelperAuth.toRegistryAuth();
+  }
 
-    final JsonNode config = MAPPER.readTree(file);
+  private String getCredentialStore(final DockerConfig config, final String registry) {
+    checkNotNull(config, "Docker config cannot be null");
+    checkNotNull(registry, "registry cannot be null");
 
-    Preconditions.checkState(config.isObject(),
-        "config file contents are not a JSON Object, instead it is a %s", config.getNodeType());
-
-    if (config.has(AUTHS_ENTRY)) {
-      final JsonNode auths = config.get(AUTHS_ENTRY);
-      Preconditions.checkState(auths.isObject(),
-          "config file contents are not a JSON Object, instead it is a %s", auths.getNodeType());
-    }
-
-    return (ObjectNode) config;
+    // Check for the registry in the credsHelpers map first.
+    // If it isn't there, default to credsStore.
+    final Map<String, String> credsHelpers = config.credsHelpers();
+    return (credsHelpers != null && credsHelpers.containsKey(registry))
+        ? credsHelpers.get(registry)
+        : config.credsStore();
   }
 }